Decouple tests
