+++ auth_provider/tests.py
@@ -1,3 +1,17 @@
-from django.test import TestCase
+from django.test import TestCase, Client
+from django.urls import reverse_lazy
 
-# Create your tests here.
+
+class ViewsTests(TestCase):
+    def setUp(self) -> None:
+        self.client = Client()
+
+    def test_login_page(self):
+        res = self.client.get(reverse_lazy('auth_provider:login'))
+        self.assertEqual(res.status_code, 200)
+        self.assertTemplateUsed('auth/login.html')
+
+    def test_signup_page(self):
+        res = self.client.get(reverse_lazy('auth_provider:signup'))
+        self.assertEqual(res.status_code, 200)
+        self.assertTemplateUsed('auth/signup.html')
